requirements.txt
googlemaps>=4.10.0
playwright>=1.40.0
apify-client>=1.6.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
openai>=1.12.0
flask>=3.0.0
requests>=2.31.0
